src/layers/shared/requirements.txt
# Note: boto3/botocore excluded - provided by Lambda runtime
pydantic[email]>=2.5.0
python-ulid>=2.2.0
structlog>=24.1.0
PyJWT>=2.8.0
cryptography>=41.0.0
httpx>=0.26.0
twilio>=9.0.0
stripe>=10.0.0
pypdf>=4.0.0
python-docx>=1.1.0
checkdmarc>=5.6.0
pydnsbl>=1.1.0
